refactor: move exporter validation
